move CreateShowing.tsx into CreateShowing folder
